pyproject.toml
[project]
name = "org-stats"
authors = [{name = "Adam Tyson", email= "code@adamltyson.com"}]
description = "Scripts to report statistics about a GitHub organisation"
readme = "README.md"
requires-python = ">=3.8.0"
dynamic = ["version"]

license = {text = "BSD-3-Clause"}

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: BSD License",
]

dependencies = ["typer", "PyGithub", "pypistats"]

[project.urls]
homepage = "https://github.com/adamltyson/org-stats"
bug_tracker = "https://github.com/adamltyson/org-stats/issues"
documentation = "https://github.com/adamltyson/org-stats"
source_code = "https://github.com/adamltyson/org-stats"
user_support = "https://github.com/adamltyson/org-stats/issues"

[project.optional-dependencies]
dev = [
	"pytest",
	"pytest-cov",
	"coverage",
	"tox",
	"black",
   	"isort",
    "mypy",
	"pre-commit",
	"flake8",
 	"setuptools_scm",
]



[build-system]
requires = [
    "setuptools>=45",
    "wheel",
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["org_stats*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
addopts = "--cov=org_stats"

[tool.black]
target-version = ['py38', 'py39', 'py310']
skip-string-normalization = false
line-length = 79
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | examples
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 79

[tool.setuptools_scm]

[tool.check-manifest]
ignore = [
  "*.yaml",
  ".bumpversion.cfg",
  "tox.ini",
  "tests/*",
  "tests/test_unit/*",
  "tests/test_integration/*",
  ".flake8"
]
